requirements.txt
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
black==22.1.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
debugpy==1.5.1
decorator==5.1.1
dnspython==2.2.0
entrypoints==0.4
executing==0.8.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.9.0
ipython==8.0.1
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.1
jmespath==0.10.0
jupyter-client==7.1.2
jupyter-core==4.9.1
lxml==4.7.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
parsel==1.6.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
priority==1.3.0
prompt-toolkit==3.0.27
Protego==0.1.16
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pymongo==4.0.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==303
pyzmq==22.3.0
queuelib==1.6.2
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
stack-data==0.1.4
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.4.0
